fix: cast pytorch benchmark outputs before numpy

+++ scripts/bench_pytorch.py
@@ -100,7 +100,12 @@ def embed_texts(model, tokenizer, texts: list[dict[str, str]]) -> list[list[floa
     )
     with torch.no_grad():
         output = model(**encoded)
-    pooled = last_token_pool(output.last_hidden_state, encoded["attention_mask"]).cpu().numpy()
+    pooled = (
+        last_token_pool(output.last_hidden_state, encoded["attention_mask"])
+        .to(dtype=torch.float32)
+        .cpu()
+        .numpy()
+    )
     normalized = truncate_and_normalize(pooled)
     return normalized.tolist()
